+++ example/main.dart
@@ -1,0 +1,6 @@
+import 'package:humanize/humanize.dart' as humanize;
+
+void main() {
+  print(humanize.appNumber(1));
+  print(humanize.ordinal(2));
+}