+++ CHANGELOG.md
@@ -1,3 +1,8 @@
-## [0.0.1] - TODO: Add release date.
 
-* TODO: Describe initial release.
+## [0.1.0] - 13 February 2020.
+
+* ordinal 
+    - Convert an integer to its ordinal as a string. 1 is '1st', 2 is '2nd', 3 is '3rd', etc. works for any integer.
+
+* appNumber
+    - For numbers 1-9, return the number spelled out. Otherwise, return the number.